requirements.txt
streamlit
requests
torch
nltk
transformers
pymilvus
pyyaml
numpy

